requirements.txt
streamlit
google-generativeai
python-dotenv
chromadb
langchain
PyPDF2
